Add endian.h shim for byteswapping function support on OS X.
There are various version of this header shim floating around online.
This one is minimalistic and only includes macros for the two functions
actually used in the code.

CoreFoundation also provides functions for handling endianness
conversions, which may or may not be more future proof than the current
include. However, including CoreFoundation causes the code to be treated
as objective-c, so using it has been avoided.